Introduce control message for clarity, change closed -> stopped
This package is for a single stream. If there is a single stream
between sender and receiver then it does not make sense to close
it.

Scenario 1: sender and receiver can agree to open the stream
again. If the application permits it.

Scenario 2: if the receiver can manage multiple streams then they
can agree to open a new stream. But managing multiple streams
happens outside this package. So whether a stream is closed for
good is a fact that is handled outside this package.