replace paused by setting head to null
In paused state the head is essentially undefined because
whenever we come out of paused we have to reset the head.
Therefore it seems safer to encode the paused state by setting
head to null. Then the code will trap if we ever forget to reset
the head and we would catch that bug.