Remove buffer.deleteTo(start)
Possible bug: two calls are made in a row, i.e. the second call is made before
the first one returns. (Note: We cannot currently simulate this with a test.)
The  first one returns with canister_reject. The second returns with #stopped
(aka timeout, formerly #err).  Then the old code would call
buffer.deleteTo(start) where start is the start value of the second call. So
we would incorrectly skip over the the first chunk which wasn't successful.