Reintroduce buffer.deleteTo(start)
The previous commit message did not describe a bug. The scenario
described is not possible because the StreamReceiver tests for a
matching chunk start index _before_ it tests for a timeout. In
other words a reject message with "broken pipe" is issued _before_
a #stopped control message is returned. Therefore, the scenario
where the first call return reject and the second returns #stopped
isn't possible. If a call returns #stopped then all previous calls
must eventually return #ok. Strictly speaking omitting the
buffer.deleteTo(start) isn't wrong because the previous calls will
return with #ok and the last one of those will call buffer.deleteTo
with the same value. However, we still re-introduce buffer.deleteTo
in this commit for the following two reasons:

1) It is consistent with how #ok messages are handled. We call
buffer.deleteTo as soon as possible, not caring about the order of
responses.

2) If the outside code observes that a stream is stopped with
isStopped() then it wants to be able to see immediately at with
index it was stopped. It does not want to wait indefinitely for
other open calls to come back. It can do that with received()
function if we re-introduce buffer.deleteTo().