sendChunk back to return (), but can throw. New status() function
The caller should inspect the stream status before calling sendChunkuser, not
after. Returning information about the stream status or the success of the
chunk from sendChunk has little value for the caller. The reason is that
sendChunk has to wait for the response from the receiver before it can itself
respond. But the responses from the receiver can arrive out of order, thus
multiple calls to sendChunk can also respond out of order. The whole point of
the StreamSender class it to hide those kind of details from the user. It would
be a challenge for the user to handle the responses correctly.

A better approach is to inspect the status of the stream before calling
sendChunk (or not at all). For that purpose this commit provides the option to
call the synchronous status() function.