Add assertions and shutdown state in the Sender
By doing so we enforce the Receiver to adhere to the protocol.
The Sender shuts down if it receives responses that are not allowed
by the protocol, which could happen if the Receiver is corrupted.

For example, receiving gap for one chunk and ok for a subsequent chunk is
invalid, regardless of the order in which the responses are received.
To catch that we changed `head` back to non-optional. The head value
now moves backwards as #gap or error responses come in.