Lift metrics to actor top-level
The PT.PromTracker class `metrics` has to live on the actor
top-level. The reason is that the actor may track many more values
besides the ones from the streams package. And all values have to
live in the same PromTracker class because there is only one http
endpoint providing them all.